requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
supabase==2.6.0
openai==1.35.7
tweepy==4.14.0
google-api-python-client==2.134.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
